play: Extract interactions
